Eliminate windows warning
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@951502 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/include/qpid/framing/FieldValue.h
@@ -123,7 +123,7 @@ template <>
 inline bool FieldValue::convertsTo<std::string>() const { return data->convertsToString(); }
 
 template <>
-inline int FieldValue::get<int>() const { return data->getInt(); }
+inline int FieldValue::get<int>() const { return static_cast<int>(data->getInt()); }
 
 template <>
 inline int64_t FieldValue::get<int64_t>() const { return data->getInt(); }
